+++ src/views/components/SearchWidget/SearchWidget.tsx
@@ -1,10 +1,9 @@
-import { RootContext } from '@data/context'
-import { Game, Genre, Publisher, Tag } from '@data/types'
+import { Genre, Publisher, Tag } from '@data/types'
 import { GenreService, PublisherService, TagService } from '@src/services'
-import * as React from 'react'
-import { useContext, useEffect, useState } from 'react'
-import { Form, Link, useNavigate } from 'react-router-dom'
+import { useEffect, useState } from 'react'
+import { Link } from 'react-router-dom'
 import style from './SearchWidget.module.scss'
+import { SearchForm } from '@views/components/SearchForm/SearchForm.tsx'
 
 
 function SearchWidget() {
@@ -30,51 +29,6 @@ function SearchWidget() {
 			<SearchForm/>
 			<SearchNavbar {...searchNavbarProps} />
 		</aside>
-	)
-}
-
-function SearchForm() {
-	const { apiMiddleware, setGames } = useContext(RootContext)
-
-	const navigator = useNavigate()
-
-	const search = (e: React.FormEvent<HTMLFormElement>) => {
-		e.preventDefault()
-
-		const data      = new FormData(e.target as HTMLFormElement)
-		const newSearch = data.get('search_content') as string
-
-		if (! newSearch) return
-
-		apiMiddleware?.getAll('games', { search: newSearch })
-			.then(apiData => setGames(apiData as Game[]))
-		navigator('/search')
-
-		setTimeout(() => window.location.assign('#search-header'), 200)
-	}
-
-	return (
-		<Form method="get"
-			  className={style.SearchForm}
-			  action="/search"
-			  onSubmit={search}
-		>
-			<label className={style.Label}
-				   htmlFor="search_content"
-			>Keyword</label>
-
-			<input className={style.Search}
-				   type="search"
-				   id="search_content"
-				   name="search_content"
-			/>
-
-			<button className={`primary ${style.Submit}`}
-					type="submit"
-			>
-				Search
-			</button>
-		</Form>
 	)
 }
